+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 11.2.0 (XX.XX.2021)
+
+### Features
+
+-   **terra-form**
+    -   form type **multi select**: replaced deprecated `tc-checkbox-group` with a `mat-select` variant in the multiple mode.
+
 # 11.1.0 (23.03.2021)
 
 ### Features